Revert "refactor: add createPreloadHandlerRef"
This reverts commit 9caa5adffd516c7f0bb6d76d838d31ae70ce3f32.
