+++ examples/spa/src/routes/pokemon/layout.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { NavLink, Outlet, useNavigate, useNavigation } from "react-router-dom";
-import { createPreloadHandlerRef } from "../../utils/preload";
 
 export function Component() {
   const navigation = useNavigation();
@@ -24,7 +23,7 @@ export function Component() {
               <NavLink
                 to={`/pokemon/${v}`}
                 style={(x) => (x.isActive ? { fontWeight: "bold" } : {})}
-                ref={createPreloadHandlerRef()}
+                data-prefetch
               >
                 {v}
               </NavLink>
